Few changes to A*
I'm honestly not sure how this compares to, say, bug pathfinding. But for the
tricky test map I'm running on, it completes the maze in about 600 turns
(fewer if it had more map info).

Still need to add alternative directions (since in the real world, allied
robots will likely be in the way).

Also ideally just one robot would do this (ideally a scout, for the large
vision and bytecodes), so we should add some follow-the-leader broadcast
semantics.