Added a turret bot similar to NP-compete's
It picks an arbitrary archon and sieges up next to it. The siege ring expands slowly if it get's filled. This beats dk007 on every released map, except ones where the initial archon is stuck in a corner (TODO: detect that case and change the numbers when that happens).
On some maps, dk007 appears to have total map control, but just loses due to archon health. A better dk007 might detect when it has map control, and then swarm the enemy siege line.

This package also revamps the pathfinding interface, maybe that'll be useful later.

Also a small change to archon retreat semantics and class naming.