Slight change to avoid overcrowding.
This lets us make more units, but also makes us more exposed
sometimes (for example, the old version of dk007 beats this
version on farm, but that's because it built less units (due
to overcrowding) and as a result there was less zombie
fodder on the deciding spawn wave). I'm not 100% sure if
this is a good commit, but it should be good in theory.